smart-agriculture-fixed/requirements.txt
Flask==3.0.0
Flask-Session==0.6.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
scikit-learn==1.3.2
tensorflow==2.15.0
pandas==2.1.4
numpy==1.26.2
Pillow==10.1.0
joblib==1.3.2
python-dotenv==1.0.0
Flask-Migrate==4.0.5
bcrypt==4.1.2
